Create SIL.FwBuildTasks.Lib nuget package
The regular `SIL.FwBuildTasks` package has the `SIL.FwBuildTasks.dll`
in `tools`. This is desirable so that it doesn't get copied to the
output directory. However, this makes it impossible for clients to
access the classes. Therefore this change adds the
`SIL.FwBuildTasks.Lib` nuget package which puts the assembly in `lib`.
To access the tasks the `SIL.FwBuildTasks` nuget package should be
installed. To create additional tasks that derive from FwBuildTasks
`SIL.FwBuildTasks.Lib` should be used.